Allow for real big matrices

+++ src/cells/AtCell.h
@@ -35,9 +35,9 @@ public:
   std::unique_ptr<Cell> Copy(GroupCell *group) const override;
   const CellTypeInfo &GetInfo() override;
 
-  int GetInnerCellCount() const override { return 2; }
+  size_t GetInnerCellCount() const override { return 2; }
   // cppcheck-suppress objectIndex
-  Cell *GetInnerCell(int index) const override { return (&m_baseCell)[index].get(); }
+  Cell *GetInnerCell(size_t index) const override { return (&m_baseCell)[index].get(); }
 
   void Recalculate(AFontSize fontsize) override;
