+++ src/cells/IntervalCell.h
@@ -55,9 +55,9 @@ public:
   virtual std::unique_ptr<Cell> Copy(GroupCell *group) const override;
   virtual const CellTypeInfo &GetInfo() override;
 
-  int GetInnerCellCount() const override { return 8; }
+  size_t GetInnerCellCount() const override { return 8; }
   // cppcheck-suppress objectIndex
-  Cell *GetInnerCell(int index) const override { return (&m_open)[index].get(); }
+  Cell *GetInnerCell(size_t index) const override { return (&m_open)[index].get(); }
 
   void Recalculate(AFontSize fontsize) override;
