+++ src/cells/LimitCell.h
@@ -47,9 +47,9 @@ public:
   std::unique_ptr<Cell> Copy(GroupCell *group) const override;
   const CellTypeInfo &GetInfo() override;
 
-  int GetInnerCellCount() const override { return 6; }
+  size_t GetInnerCellCount() const override { return 6; }
   // cppcheck-suppress objectIndex
-  Cell *GetInnerCell(int index) const override { return (&m_name)[index].get(); }
+  Cell *GetInnerCell(size_t index) const override { return (&m_name)[index].get(); }
 
   void Recalculate(AFontSize fontsize) override;
